client: rename redis-bytes-ref to redis-bytes-get
Initially, I thought it'd make sense to use ref and have the ops be
similar to dicts and hash tables, but the operation is different
enough from a -ref that it makes sense to use a different name.  Plus,
it didn't make sense for redis-bytes-ref and redis-list-ref to have
similar names, but perform completely different operations.