client: connect lazily, fix auth and db selection on reconnect
If an error is raised during connect, that can affect the performance
of the connection pool, so avoid that by ensuring that make-redis can
never raise an exception. Things were already set up so that
connections could be lazy so this change takes advantage of that.

Also, fix db and auth selection. Previously, on reconnect, the client
wouldn't select the right database or send auth commands.